наткнулся на баг reload таймера However, this does not usually work, probably because the timer runs a little slower than the CPU, and the APB bus also needs a few cycles to transfer the access to EGR  https://community.st.com/t5/stm32-mcus-embedded-software/bug-in-tim-base-setconfig-fix-tim-base-setconfig-to-block-first/m-p/754265
провёл рефакторинг и проверку установки флага в цикле
по косвенным замерам надо 45 mks, но это не точно

+++ examples/stm32_irq/Core/Src/tim.c
@@ -21,7 +21,7 @@
 #include "tim.h"
 
 /* USER CODE BEGIN 0 */
-extern uint16_t Speed;
+extern uint32_t Speed;
 extern uint32_t FastModbus_Prescaler, Arbitrage_Period, Window_Period,
 		Normal_Prescaler, Normal_Period;
 /* USER CODE END 0 */
